requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
coverage==7.10.6
cryptography==45.0.7
Django==5.2.5
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-jwt==1.19.2
exceptiongroup==1.3.0
execnet==2.1.1
factory_boy==3.3.3
Faker==37.6.0
freezegun==1.5.5
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
py-cpuinfo==9.0.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-benchmark==5.1.0
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-mock==3.15.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
responses==0.25.8
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
